+++ component/loki/process/stages/extensions.go
@@ -5,17 +5,20 @@ package stages
 // new code without being able to slowly review, examine and test them.
 
 import (
+	"fmt"
 	"strings"
 
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
+	"github.com/grafana/river"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/common/model"
+
+	"github.com/grafana/loki/pkg/util/flagext"
 )
 
 const (
-	RFC3339Nano         = "RFC3339Nano"
-	MaxPartialLinesSize = 100 // MaxPartialLinesSize is the max buffer size to hold partial lines when parsing the CRI stage format.lines.
+	RFC3339Nano = "RFC3339Nano"
 )
 
 // DockerConfig is an empty struct that is used to enable a pre-defined
@@ -24,7 +27,37 @@ type DockerConfig struct{}
 
 // CRIConfig is an empty struct that is used to enable a pre-defined pipeline
 // for decoding entries that are using the CRI logging format.
-type CRIConfig struct{}
+type CRIConfig struct {
+	MaxPartialLines            int              `river:"max_partial_lines,attr,optional"`
+	MaxPartialLineSize         flagext.ByteSize `river:"max_partial_line_size,attr,optional"`
+	MaxPartialLineSizeTruncate bool             `river:"max_partial_line_size_truncate,attr,optional"`
+}
+
+var (
+	_ river.Defaulter = (*CRIConfig)(nil)
+	_ river.Validator = (*CRIConfig)(nil)
+)
+
+// DefaultCRIConfig applies the default values on
+var DefaultCRIConfig = CRIConfig{
+	MaxPartialLines:            100,
+	MaxPartialLineSize:         0,
+	MaxPartialLineSizeTruncate: false,
+}
+
+// SetToDefault implements river.Defaulter.
+func (args *CRIConfig) SetToDefault() {
+	*args = DefaultCRIConfig
+}
+
+// Validate implements river.Validator.
+func (args *CRIConfig) Validate() error {
+	if args.MaxPartialLines <= 0 {
+		return fmt.Errorf("max_partial_lines must be greater than 0")
+	}
+
+	return nil
+}
 
 // NewDocker creates a predefined pipeline for parsing entries in the Docker
 // json log format.
@@ -61,17 +94,19 @@ func NewDocker(logger log.Logger, registerer prometheus.Registerer) (Stage, erro
 
 type cri struct {
 	// bounded buffer for CRI-O Partial logs lines (identified with tag `P` till we reach first `F`)
-	partialLines    map[model.Fingerprint]Entry
-	maxPartialLines int
-	base            *Pipeline
+	partialLines map[model.Fingerprint]Entry
+	cfg          CRIConfig
+	base         *Pipeline
 }
+
+var _ Stage = (*cri)(nil)
 
 // Name implement the Stage interface.
 func (c *cri) Name() string {
 	return "cri"
 }
 
-// Run implements Stage interface
+// implements Stage interface
 func (c *cri) Run(entry chan Entry) chan Entry {
 	entry = c.base.Run(entry)
 
@@ -80,16 +115,17 @@ func (c *cri) Run(entry chan Entry) chan Entry {
 
 		// We received partial-line (tag: "P")
 		if e.Extracted["flags"] == "P" {
-			if len(c.partialLines) > c.maxPartialLines {
+			if len(c.partialLines) >= c.cfg.MaxPartialLines {
 				// Merge existing partialLines
 				entries := make([]Entry, 0, len(c.partialLines))
 				for _, v := range c.partialLines {
 					entries = append(entries, v)
 				}
 
-				level.Warn(c.base.logger).Log("msg", "cri stage: partial lines upperbound exceeded. merging it to single line", "threshold", MaxPartialLinesSize)
+				level.Warn(c.base.logger).Log("msg", "cri stage: partial lines upperbound exceeded. merging it to single line", "threshold", c.cfg.MaxPartialLines)
 
-				c.partialLines = make(map[model.Fingerprint]Entry)
+				c.partialLines = make(map[model.Fingerprint]Entry, c.cfg.MaxPartialLines)
+				c.ensureTruncateIfRequired(&e)
 				c.partialLines[fingerprint] = e
 
 				return entries, false
@@ -97,8 +133,12 @@ func (c *cri) Run(entry chan Entry) chan Entry {
 
 			prev, ok := c.partialLines[fingerprint]
 			if ok {
-				e.Line = strings.Join([]string{prev.Line, e.Line}, "")
+				var builder strings.Builder
+				builder.WriteString(prev.Line)
+				builder.WriteString(e.Line)
+				e.Line = builder.String()
 			}
+			c.ensureTruncateIfRequired(&e)
 			c.partialLines[fingerprint] = e
 
 			return []Entry{e}, true // it's a partial-line so skip it.
@@ -109,7 +149,11 @@ func (c *cri) Run(entry chan Entry) chan Entry {
 		// 2. Else just return the full line.
 		prev, ok := c.partialLines[fingerprint]
 		if ok {
-			e.Line = strings.Join([]string{prev.Line, e.Line}, "")
+			var builder strings.Builder
+			builder.WriteString(prev.Line)
+			builder.WriteString(e.Line)
+			e.Line = builder.String()
+			c.ensureTruncateIfRequired(&e)
 			delete(c.partialLines, fingerprint)
 		}
 		return []Entry{e}, false
@@ -118,9 +162,15 @@ func (c *cri) Run(entry chan Entry) chan Entry {
 	return in
 }
 
+func (c *cri) ensureTruncateIfRequired(e *Entry) {
+	if c.cfg.MaxPartialLineSizeTruncate && len(e.Line) > c.cfg.MaxPartialLineSize.Val() {
+		e.Line = e.Line[:c.cfg.MaxPartialLineSize.Val()]
+	}
+}
+
 // NewCRI creates a predefined pipeline for parsing entries in the CRI log
 // format.
-func NewCRI(logger log.Logger, registerer prometheus.Registerer) (Stage, error) {
+func NewCRI(logger log.Logger, config CRIConfig, registerer prometheus.Registerer) (Stage, error) {
 	base := []StageConfig{
 		{
 			RegexConfig: &RegexConfig{
@@ -156,9 +206,9 @@ func NewCRI(logger log.Logger, registerer prometheus.Registerer) (Stage, error) 
 	}
 
 	c := cri{
-		maxPartialLines: MaxPartialLinesSize,
-		base:            p,
+		cfg:  config,
+		base: p,
 	}
-	c.partialLines = make(map[model.Fingerprint]Entry)
+	c.partialLines = make(map[model.Fingerprint]Entry, c.cfg.MaxPartialLines)
 	return &c, nil
 }